menambahkan konfirmasi penghapusan di read.php untuk link delate

+++ pertemuan-13/read.php
@@ -47,7 +47,7 @@
       <td><?= $i++ ?></td>
       <td>
         <a href="edit.php?cid=<?= (int)$row['cid']; ?>">Edit</a>
-         <a href="proses_delete.php?cid=<?= (int)$row['cid']; ?>">delete</a>
+        <a onclick="return confirm('HAPUS?')" href="proses_delete.php?cid=<?= (int)$row['cid']; ?>">delete</a>
     </td>
       <td><?= $row['cid']; ?></td>
       <td><?= htmlspecialchars($row['cnama']); ?></td>
